renamed computed helpers
