+++ addon/computeds/moment-ago.js
@@ -1,13 +1,13 @@
 import Ember from 'ember';
 import moment from 'moment';
-import { descriptorFor } from './moment';
+import { descriptorFor } from './moment-format';
 
 var get = Ember.get;
 var emberComputed = Ember.computed;
 
 export default function computedAgo(date, maybeInputFormat) {
   var args = [date];
-  
+
   var computed = emberComputed(date, function () {
     var momentArgs, desc, input;
     momentArgs = [get(this, date)];
